macOS: extract surface environment and focus keys into SwiftUI+Extension.swift
The SwiftUI environment and focused-value plumbing for surfaces was appended
to the bottom of SurfaceView.swift, outside the `extension Ghostty` block it
had nothing to do with. Move it to Ghostty/SwiftUI+Extension.swift.

Pure code motion; the keys and their definitions are unchanged.
