+++ macos/Sources/Ghostty/Surface View/SurfaceView.swift
@@ -751,69 +751,3 @@ extension Ghostty {
         }
     }
 }
-
-// MARK: Surface Environment Keys
-
-private struct GhosttySurfaceViewKey: EnvironmentKey {
-    static let defaultValue: Ghostty.SurfaceView? = nil
-}
-
-private struct GhosttyLastFocusedSurfaceKey: EnvironmentKey {
-    /// Optional read-only last-focused surface reference. If a surface view is currently focused this
-    /// is equal to the currently focused surface.
-    static let defaultValue: Weak<Ghostty.SurfaceView>? = nil
-}
-
-extension EnvironmentValues {
-    var ghosttySurfaceView: Ghostty.SurfaceView? {
-        get { self[GhosttySurfaceViewKey.self] }
-        set { self[GhosttySurfaceViewKey.self] = newValue }
-    }
-
-    var ghosttyLastFocusedSurface: Weak<Ghostty.SurfaceView>? {
-        get { self[GhosttyLastFocusedSurfaceKey.self] }
-        set { self[GhosttyLastFocusedSurfaceKey.self] = newValue }
-    }
-}
-
-extension View {
-    func ghosttySurfaceView(_ surfaceView: Ghostty.SurfaceView?) -> some View {
-        environment(\.ghosttySurfaceView, surfaceView)
-    }
-
-    /// The most recently focused surface (can be currently focused if the surface is currently focused).
-    func ghosttyLastFocusedSurface(_ surfaceView: Weak<Ghostty.SurfaceView>?) -> some View {
-        environment(\.ghosttyLastFocusedSurface, surfaceView)
-    }
-}
-
-// MARK: Surface Focus Keys
-
-extension FocusedValues {
-    var ghosttySurfaceView: Ghostty.SurfaceView? {
-        get { self[FocusedGhosttySurface.self] }
-        set { self[FocusedGhosttySurface.self] = newValue }
-    }
-
-    struct FocusedGhosttySurface: FocusedValueKey {
-        typealias Value = Ghostty.SurfaceView
-    }
-
-    var ghosttySurfacePwd: String? {
-        get { self[FocusedGhosttySurfacePwd.self] }
-        set { self[FocusedGhosttySurfacePwd.self] = newValue }
-    }
-
-    struct FocusedGhosttySurfacePwd: FocusedValueKey {
-        typealias Value = String
-    }
-
-    var ghosttySurfaceCellSize: CGSize? {
-        get { self[FocusedGhosttySurfaceCellSize.self] }
-        set { self[FocusedGhosttySurfaceCellSize.self] = newValue }
-    }
-
-    struct FocusedGhosttySurfaceCellSize: FocusedValueKey {
-        typealias Value = CGSize
-    }
-}
